+++ packages/seq-quality/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+
+## [0.2.0] - 2026-07-29
+
+### Changed
+- Dependency ranges updated for `@bioscript/seq-utils@^0.3.0` and `@bioscript/seq-stream@^0.2.0`
+
 ## [0.1.0] - 2026-07-29
 
 ### Added