+++ packages/seq-search/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+
+## [0.2.0] - 2026-07-29
+
+### Fixed
+- Palindromic restriction sites are reported once (strand dedupe)
+
 ## [0.1.0] - 2026-07-29
 
 ### Added